Correcao do pagamento por cartao no superpay
