+++ src/test/java/com/cosmeticos/controller/PaymentControllerTests.java
@@ -7,6 +7,7 @@ import com.cosmeticos.commons.OrderResponseBody;
 import com.cosmeticos.model.*;
 import com.cosmeticos.payment.superpay.client.rest.model.RetornoTransacao;
 import com.cosmeticos.repository.*;
+import com.cosmeticos.service.PaymentService;
 import com.cosmeticos.validation.OrderValidationException;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.junit.Assert;
@@ -45,7 +46,7 @@ public class PaymentControllerTests {
     private CustomerRepository customerRepository;
 
     @Autowired
-    private PaymentController paymentController;
+    private PaymentService paymentService;
 
     @Autowired
     AddressRepository addressRepository;
@@ -370,7 +371,7 @@ public class PaymentControllerTests {
 
         /************ FIM DAS PRE_CONDICOES **********************************/
 
-        Optional<RetornoTransacao> retornoTransacao = paymentController.sendRequest(order);
+        Optional<RetornoTransacao> retornoTransacao = paymentService.sendRequest(order);
 
         Assert.assertNotNull(retornoTransacao.isPresent());
         Assert.assertNotNull(retornoTransacao.get().getAutorizacao());
@@ -386,7 +387,7 @@ public class PaymentControllerTests {
         //CERTIFIQUE-SE QUE A ORDER ABAIXO ESTA COM O STATUS READY2CHARGE, CASO CONTRARIO RETORNARA UM ERRO!
         Order order = orderRepository.findOne(14L);
 
-        Boolean capturaTransacao = paymentController.validatePaymentStatusAndSendCapture(order);
+        Boolean capturaTransacao = paymentService.validatePaymentStatusAndSendCapture(order);
 
         Assert.assertNotNull(capturaTransacao);
         Assert.assertEquals(true, capturaTransacao);
